+++ public/components/commentCard.html
@@ -23,6 +23,15 @@
       <div class="comment__content"></div>
     </div>
 
+    <!-- 내장 답글 폼 (기본 숨김) -->
+    <form class="reply-form hidden" novalidate>
+      <textarea rows="2" placeholder="답글을 입력하세요…"></textarea>
+      <div class="reply-actions">
+        <button type="submit" class="btn btn--small btn--primary">등록</button>
+        <button type="button" class="btn btn--small btn--ghost" data-action="reply-cancel">취소</button>
+      </div>
+    </form>
+
     <!-- 대댓글이 들어갈 영역 -->
     <div class="replies"></div>
   </article>
